refactor(app): move cal warning
