+++ app/src/organisms/Devices/InstrumentsAndModules.tsx
@@ -26,7 +26,7 @@ import { useCurrentRunId } from '/app/resources/runs'
 import { useIsFlex } from '/app/redux-resources/robots'
 import { ModuleCard } from '../ModuleCard'
 import { useIsRobotViewable, useRunStatuses } from './hooks'
-import { getShowPipetteCalibrationWarning } from './utils'
+import { getShowPipetteCalibrationWarning } from '/app/transformations/instruments'
 import { PipetteCard } from './PipetteCard'
 import { FlexPipetteCard } from './PipetteCard/FlexPipetteCard'
 import { GripperCard } from '../GripperCard'
