+++ app/src/organisms/RobotSettingsCalibration/RobotSettingsPipetteOffsetCalibration.tsx
@@ -15,7 +15,7 @@ import {
   usePipetteOffsetCalibrations,
 } from '../Devices/hooks'
 import { useIsFlex } from '/app/redux-resources/robots'
-import { getShowPipetteCalibrationWarning } from '../Devices/utils'
+import { getShowPipetteCalibrationWarning } from '/app/transformations/instruments'
 import { PipetteRecalibrationWarning } from '../Devices/PipetteCard/PipetteRecalibrationWarning'
 import { PipetteOffsetCalibrationItems } from './CalibrationDetails/PipetteOffsetCalibrationItems'
